weather_app_api/requirements.txt
requests
python-dotenv
python-tkinter
python-os